continue last upgrade.

+++ compatible_edition/include/st_asio_wrapper_base.h
@@ -345,7 +345,7 @@ public:
 #ifdef _MSC_VER
 		os.rdbuf()->sgetn(buff, buff_len);
 #endif
-		len = os.tellp();
+		len = (size_t) os.tellp();
 		if (len >= buff_len)
 			*boost::next(buff, buff_len - 1) = '\0';
 		else
